+++ components/goal-hierarchy/index.ts
@@ -1,3 +1,5 @@
 export { GoalHierarchyView } from './GoalHierarchyView';
 export { GoalLevelBlock } from './GoalLevelBlock';
+export { DayGoalTextarea } from './DayGoalTextarea';
+export { DayGoalShortcutsHelp } from './DayGoalShortcutsHelp';
 export { DayColumn } from './DayColumn';